+++ dcm4chee/dcm4chee-arr/trunk/dcm4chee-arr-seam-ejb/src/main/java/org/dcm4chee/arr/seam/ejb/AuditRecordEntry.java
@@ -94,7 +94,7 @@ public class AuditRecordEntry {
         return XSLTUtils.render(XSLTUtils.OBJECT, auditRecord.getXmldata());
     }
 
-    public int getPk() {
+    public long getPk() {
         return auditRecord.getPk();
     }
